Organize app into modules
