+++ ACTTV/__init__.py
@@ -1,2 +1,3 @@
 # This package marker file allows Assetto Corsa to recognize the app
-# directory as a Python package.  It can be empty.
+# directory as a Python package. It can be empty.
+
