+++ ACTTV/config.py
@@ -1,0 +1,11 @@
+"""Configuration constants for the ACTTV app."""
+
+# Application name
+APP_NAME = "ACTTV"
+
+# --- Config ---
+BASE_INTERVAL = 10.0     # base seconds between switches
+JITTER = 3.0             # random +/- seconds
+MIN_CARS_REQUIRED = 1
+NEAR_RADIUS_M = 20.0     # proximity radius (meters) for "battle" scoring
+
