+++ ACTTV/state.py
@@ -1,0 +1,12 @@
+"""Shared state for the ACTTV app."""
+
+# --- UI / state ---
+app_window = None
+status_label = None
+camera_label = None
+toggle_button = None
+
+next_switch_time = 0.0
+current_car_id = -1
+enabled = True
+
